improved(display): transformed text appears more crisp
fixed broken text on line geometry

+++ packages/display/src/displays/webgl/buffer/templates/TextBuffer.ts
@@ -36,7 +36,7 @@ export class TextBuffer extends TemplateBuffer {
                 size: 3
             },
             a_texcoord: {
-                data: new FlexArray(Float32Array),
+                data: new FlexArray(Uint16Array),
                 size: 2
             }
         };
